% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_arrow}
\alias{align_arrow}
\title{Align a highlighted region's assignment operators.}
\usage{
align_arrow()
}
\value{
Aligns the single assignment operators (\code{<-}) within a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
\examples{
\dontrun{
# You'll likely never use remedy functions directly
# But rather map them to keyboard shortcut
align_arrow()
}
}
