% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_equal}
\alias{align_equal}
\title{Align a highlighted region's assignment operators.}
\usage{
align_equal()
}
\value{
Aligns the equal sign assignment operators (\code{=}) within a
highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
\examples{
\dontrun{
remedy_example(
    c( "# Align equal signs",
    "a = 12", 
    "aaa = 13"), 
    align_equal
    )
}
}
