% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignAssign}
\alias{alignAssign}
\title{Align a highlighted region's assignment operators.}
\usage{
alignAssign()
}
\value{
Aligns the single assignment operators (\code{<-}) within a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
