% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignAssign2}
\alias{alignAssign2}
\title{Align a highlighted region's assignment operators.}
\usage{
alignAssign2()
}
\value{
Aligns the equal sign assignment operators (\code{=}) within a
highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
