% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtickr.R
\name{backtickr}
\alias{backtickr}
\title{Backtick}
\usage{
backtickr()
}
\value{
a backticked text selection
}
\description{
Backtick selected word(s)
}
\examples{
\dontrun{
remedy_example(
    c( "to_backtick"), 
    backtickr
    )
}
}
