% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockquoter.R
\name{blockquoter}
\alias{blockquoter}
\title{Convert to blockquote}
\usage{
blockquoter()
}
\value{
a markdown blockquote
}
\description{
Turn the selected text into a blockquote.
}
\examples{
\dontrun{
remedy_example(
    c( "to_blockquote"), 
    blockquoter
    )
}
}
