% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{Interactivley Create Rmarkdown Chunks}
\usage{
chunkify()
}
\description{
Addin that splits full documents into Rmarkdown chunks or
wraps highlighted text in an active RStudio document editor in a Rmarkdown chunk.
}
\details{
Setting \code{\link{remedy_opts}}$set(list(full_doc=TRUE)) chunky will convert entire document by
\code{\link{remedy_opts}}$get('token'). The default token is the output from a \code{\link[knitr]{purl}}
conversion (## ----chunk name,chunk options----).

Setting \code{\link{remedy_opts}}$set(list(full_doc=FALSE)), user highlights text and chunky will
 wrap a new Rmarkdown chunk around it, utilizing the other options in \code{\link{remedy_opts}}.
}
