% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunknamer.R
\name{chunknamer}
\alias{chunknamer}
\title{Interactively Add Names to Chunks}
\usage{
chunknamer()
}
\description{
Addin that add in bulk a names to unnamed chunks in a 
Rmarkdown document in the source editor.
}
\details{
By default the addin will use "remedy" as the stem of the chunk names. this an be changed using
\code{\link{remedy_opts}}$set(name='ANOTHER NAME'). The names are then added in sequential order using
an %02d naming scheme.
}
