% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunksplitr.R
\name{chunksplitr}
\alias{chunksplitr}
\title{Split Chunk}
\usage{
chunksplitr()
}
\description{
Split a chunk into two chunks
}
\details{
Will split below row of the current cursor position or under the highlighted text.

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/chunksplit.gif?raw=true">
}}
}
