% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnoter.R
\name{footnoter}
\alias{footnoter}
\title{Footnote}
\usage{
footnoter()
}
\value{
footnote text (in markdown)
}
\description{
Create a footnote from selected text.
}
\examples{
\dontrun{
remedy_example(
    c( "# Create a footnote"), 
    footnoter
    )
}
}
