% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionwrapr.R
\name{functionwrapr}
\alias{functionwrapr}
\title{Function Wrap}
\usage{
functionwrapr(prefix = "(", postfix = ")")
}
\value{
text wrapped with parentheses with cursor set at leading parenthesis
}
\description{
Wrap selected text in a function call
}
\examples{
\dontrun{
remedy_example(
    c("wrapped in function call"), 
    functionwrapr
    )
}


}
