% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_link}
\alias{is_link}
\title{Check if text is a relative link or URL}
\usage{
is_link(text)
}
\arguments{
\item{text}{A character string}
}
\value{
TRUE or FALSE
}
\description{
Check if text is a relative link or URL
}
\examples{
is_link("www.google.com")
is_link("figs/plot.png")
}
