% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexr.R
\name{latexr}
\alias{latexr}
\title{LaTeX}
\usage{
latexr()
}
\value{
inline LaTeX text (in markdown)
}
\description{
Convert the selected text in inline LaTeX.

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/latex.gif?raw=true">
}}
}
\examples{
\dontrun{
remedy_example( 
    c( "Latex stuffs"), 
    latexr
    )
}

}
