% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryr.R
\name{libraryr}
\alias{libraryr}
\title{Interactively Add Libraries Chunk to Rmarkdown Document}
\usage{
libraryr()
}
\description{
Addin that scans the Rmarkdown Documents and adds the relevant
library calls to a chunk at top of the file.
}
