% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listr.R
\name{listr}
\alias{listr}
\alias{olistr}
\title{Convert to unordered list}
\usage{
listr()

olistr()
}
\value{
an unordered markdown list

an ordered markdown list
}
\description{
Turn the selected text into an unordered list.

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/list.gif?raw=true">
}}

Turn the selected text into an ordered list.
}
