% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mover.R
\name{rightr}
\alias{rightr}
\title{Mover}
\usage{
rightr()
}
\value{
a clone of the line
}
\description{
Copy a selected text or the current line

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/right.gif?raw=true">
}}
}
