% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referr.R
\name{id_ref}
\alias{id_ref}
\title{urlr}
\usage{
id_ref()
}
\value{
A referenced content
}
\description{
Convert selected text into a markdown ref.
}
\note{
Only the three most common reference types are implemented: id, fig, and tab.
}
\examples{
\dontrun{
remedy_example( 
    "pouet", 
    id_ref
    )
}

}
