% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remedy-package.R
\docType{package}
\name{remedy-package}
\alias{remedy}
\alias{remedy-package}
\title{remedy: 'RStudio' Addins to Simplify 'Markdown' Writing}
\description{
An 'RStudio' addin providing shortcuts for writing in 'Markdown'. This package provides a series of 
             functions that allow the user to be more efficient when using 'Markdown'. For example, you can select 
             a word, and put it in bold or in italics, or change the alignment of elements inside you Rmd. The idea
             is to map all the functionalities from 'remedy' on keyboard shortcuts, so that it provides an interface 
             close to what you can find in any other text editor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/remedy}
  \item Report bugs at \url{https://github.com/ThinkR-open/remedy/issues}
}

}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (0000-0001-7343-1846)

Authors:
\itemize{
  \item Jonathan Sidi \email{yonicd@gmail.com} (0000-0002-4222-1819)
  \item Luke Smith \email{luke@protocolvital.info} (author of seasmith/AlignAssign)
}

Other contributors:
\itemize{
  \item Jonathan Carroll \email{rpkg@jcarroll.com.au} (0000-0002-1404-5264) [contributor]
  \item Andrzej Oleś \email{andrzej.oles@gmail.com} (0000-0003-0285-2787) [contributor]
  \item Daniel Possenriede \email{possenriede@gmail.com} (0000-0002-6738-9845) [contributor]
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
