% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMeDy.R
\name{ReMeDy}
\alias{ReMeDy}
\title{A Flexible Statistical Framework For Region-based Detection of DNA Methylation Dysregulation}
\usage{
ReMeDy(
  region_list,
  metadata,
  expVar = "Exposure",
  coVars = NULL,
  rand_effect,
  parallel = FALSE,
  verbose = FALSE,
  numCores = 1
)
}
\arguments{
\item{region_list}{A list of co-methylated regions of methylation levels with CpGs in the row and samples in the column.}

\item{metadata}{A dataframe with information on the samples such as disease status, gender and etc.}

\item{expVar}{The name of the variable on which the differential methylation, variable methylation or both will be evaluated. If the user provides no name then the metadata should have a column named as exposure which should have information on things such as disease status, treatment conditions or etc. ‘Exposure’ by default.}

\item{coVars}{The names of the covariates/confounders that needs to adjusted for in the DNA methylation analysis. 'NULL' by default}

\item{rand_effect}{This parameter specifies the random-effects structure of the model.}

\item{parallel}{If true, the analysis will be performed on multiple cores with faster runtimes.'FALSE' by default}

\item{verbose}{If true, it will print the progress report. 'FALSE' by default}

\item{numCores}{The number of cores on which the analysis will be serially performed. The user needs to specify this only when parallel = TRUE.'1' by default}
}
\value{
A list containing ReMeDy results with
p-values and adjusted p-values for mean, variance and joint effects, along
with runtime and peak memory usage.
}
\description{
ReMeDy is a region-based statistical framework that detects differentially methylated regions (DMRs), variably methylated regions (VMRs) and differentially and variably methylated regions (DVMRs).
}
\author{
Suvo Chatterjee
Siddhant Meshram
Arindam Fadikar
}
