% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{remedy_example}
\alias{remedy_example}
\alias{scratch_file}
\alias{entire_document}
\alias{individual_lines}
\alias{set_text}
\title{Functions to run examples and tests}
\usage{
remedy_example(txt, addin, mark = entire_document)

scratch_file()

entire_document()

individual_lines()

set_text(txt = "", sec, mark)
}
\arguments{
\item{txt}{character, text to place in temporary document}

\item{addin}{function, addin function to deploy}

\item{mark}{function, highlight using entire_document or individual_lines}

\item{sec}{source editor context object}
}
\value{
New document in source editor of RStudio
}
\description{
Utility functions to deploy functions and tests for addins
}
\details{
These function only will work in a RStudio IDE
}
