% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotkeys.R
\name{set_hotkeys}
\alias{set_hotkeys}
\title{Set remedy addin hotkeys}
\usage{
set_hotkeys(overide = FALSE, verbose = TRUE)
}
\arguments{
\item{overide}{boolean, set to TRUE to ignore any existing hotkeys, Default: FALSE}

\item{verbose}{boolean, set to TRUE to return a message on actions taken, Default: TRUE}
}
\description{
Helper function that sets remedy addin hotkeys to defaults
}
