% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Tabler}
\usage{
tabler()
}
\value{
a markdown table
}
\description{
Insert a table in your file.

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/table.gif?raw=true">
}}

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/table_remedy.gif?raw=true">
}}
}
\examples{
\dontrun{
remedy_example( 
    "head(iris)", 
    tabler
    )
}

}
