% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{describer}
\alias{describer}
\alias{itr}
\alias{testthatr}
\title{addins to add testthat functions around expectations}
\usage{
describer()

itr()

testthatr()
}
\description{
construct testthat test_that or describe/it layouts. Highlight script in editor and
great wrappers for testthat tests.
}
\details{
In describe addin if there is nothing highlighed then a describe and it tandem will be created.
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}},\code{\link[rstudioapi]{rstudio-documents}}
}
