% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlr.R
\name{urlr}
\alias{urlr}
\title{urlr}
\usage{
urlr()
}
\value{
a markdown link
}
\description{
Convert selected text into a link

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/url.gif?raw=true">
}}
}
\examples{
\dontrun{
remedy_example( 
    "https://www.thinkr.fr", 
    urlr
    )
}

}
