% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan.R
\name{xaringanr}
\alias{xaringanr}
\title{xaringan}
\usage{
xaringanr()
}
\value{
a template for xaringan pull left and right
}
\description{
xaringan pull left and pull right

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/xaringan.gif?raw=true">
}}
}
\examples{
\dontrun{
remedy_example( 
    "", 
    xaringanr
    )
}

}
