% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\alias{align_arrow}
\alias{align_equal}
\title{Align a highlighted region's by pattern.}
\usage{
align(pattern = remedy_opts$get("align_pattern"),
  anchor = remedy_opts$get("align_anchor"))

align_arrow(pattern = "<-", anchor = remedy_opts$get("align_anchor"))

align_equal(pattern = "=", anchor = remedy_opts$get("align_anchor"))
}
\arguments{
\item{pattern}{character, regex pattern to align on, Default: remedy_opts$get('align_pattern')}

\item{anchor}{character, align to the 'right' or 'left', Default: remedy_opts$get('align_anchor')}
}
\value{
Aligns by regex pattern within a highlighted region.
}
\description{
Align a highlighted region's by pattern.
}
\examples{
\dontrun{
remedy_example(
    c( "# Align arrows",
    "a <- 12", 
    "aaa <- 13"), 
    align_arrow
    )
    
remedy_example(
    c( "# Align equal signs",
    "a = 12", 
    "aaa = 13"), 
    align_equal
    )
}
}
