% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_arrow}
\alias{align_arrow}
\title{Align a highlighted region's assignment operators.}
\usage{
align_arrow()
}
\value{
Aligns the single assignment operators (\code{<-}) within a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
\examples{
\dontrun{
remedy_example(
    c( "# Align arrows",
    "a <- 12", 
    "aaa <- 13"), 
    align_arrow
    )
}
}
