% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunksplitr.R
\name{chunksplitr}
\alias{chunksplitr}
\title{Split Chunk}
\usage{
chunksplitr()
}
\description{
Split a chunk into two chunks
}
\details{
Will split below row of the current cursor position or under the highlighted text.
}
