% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasize.R
\name{italicsr}
\alias{italicsr}
\alias{boldr}
\alias{strikr}
\title{Emphasize}
\usage{
italicsr()

boldr()

strikr()
}
\value{
emboldened, italicized or strikethrough text (in markdown)
}
\description{
Emphasize the selected text.
}
