% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasize.R
\name{italicsr}
\alias{italicsr}
\alias{boldr}
\alias{striker}
\title{Emphasize}
\usage{
italicsr()

boldr()

striker()
}
\value{
emboldened, italicized or strikethrough text (in markdown)
}
\description{
Emphasize the selected text.
}
\examples{
\dontrun{
remedy_example(
    c( "to_italic"), 
    italicsr
    )
remedy_example(
    c( "to_bold"), 
    boldr
    )
remedy_example(
    c( "to_strike"), 
    striker
    )
}

}
