% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headr.R
\name{h1r}
\alias{h1r}
\alias{h2r}
\alias{h3r}
\alias{h4r}
\alias{h5r}
\alias{h6r}
\title{Chunckize}
\usage{
h1r()

h2r()

h3r()

h4r()

h5r()

h6r()
}
\value{
a header
}
\description{
Turn a selected text into a header.
}
