% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headr.R
\name{h1r}
\alias{h1r}
\alias{h2r}
\alias{h3r}
\alias{h4r}
\alias{h5r}
\alias{h6r}
\title{Convert to header}
\usage{
h1r()

h2r()

h3r()

h4r()

h5r()

h6r()
}
\value{
a markdown header
}
\description{
Convert selected text into a header.
}
\examples{
\dontrun{
remedy_example( c( "h1"), h1r)
remedy_example( c( "h2"), h2r)
remedy_example( c( "h3"), h3r)
remedy_example( c( "h4"), h4r)
remedy_example( c( "h5"), h5r)
remedy_example( c( "h6"), h6r)
}
}
