% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imager.R
\name{imager}
\alias{imager}
\title{Imager}
\usage{
imager()
}
\value{
a markdown image link
}
\description{
Convert the selected path into an embedded image
}
\examples{
\dontrun{
remedy_example( 
    c( "https://thinkr.fr/wp-content/uploads/2015/03/thinkR1.png"), 
    imager
    )
}
}
