% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listr.R
\name{listr}
\alias{listr}
\alias{olistr}
\title{Convert to list}
\usage{
listr()

olistr()
}
\value{
\code{listr()} returns an unordered markdown list

\code{olistr()} returns an ordered markdown list
}
\description{
Convert selected text into an (un)ordered list.
}
\examples{
\dontrun{
#unordered list
remedy_example(c('line 1','line 2'),listr)

#ordered list
remedy_example(c('line 1','line 2'),olistr) 
}
}
