% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mover.R
\name{rightr}
\alias{rightr}
\title{Mover}
\usage{
rightr()
}
\value{
a clone of the line
}
\description{
Copy a selected text or the current line to the right
}
\examples{
\dontrun{
remedy_example( 
    c( "I like to move it "), 
    rightr
    )
}

}
