% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Tabler}
\usage{
tabler()
}
\value{
a markdown table
}
\description{
Insert a table in your file.
}
\examples{
\dontrun{
remedy_example( 
    "head(iris)", 
    tabler
    )
}

}
