% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlr.R
\name{urlr}
\alias{urlr}
\title{urlr}
\usage{
urlr()
}
\value{
a markdown image link
}
\description{
Convert selected text into a url

\if{html}{
\out{
 <img src = "https://github.com/ThinkR-open/remedy/blob/master/readme_gif/url.gif?raw=true">
}}
}
