% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtuber.R
\name{youtuber}
\alias{youtuber}
\title{Addin to create markdown youtube links}
\usage{
youtuber()
}
\description{
Highlight youtube link in source editor and it will be converted
to a markdown link with an image saved in youtube.
}
\details{
the frame output is in the form of ![[Alternative text](Youtube Image)](Link to Video)
}
