% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtuber.R
\name{youtuber}
\alias{youtuber}
\title{Addin to create markdown youtube links}
\usage{
youtuber()
}
\description{
Highlight youtube link in source editor and it will be converted
to a markdown link with an image saved in youtube.
}
\details{
if remedy_opts$get('youtube_output')=='md' then the output will be of the form

 [![Alternative text](Youtube Image)](Link to Video). (useful for github readme)
 
if remedy_opts$get('youtube_output')=='html' then the output will be of the form

<iframe width="remedy_opts$get('youtube_width')" height="remedy_opts$get('youtube_height')" src="link to video" 
frameborder="0" allowfullscreen></iframe>

Default Settings:
 
remedy_opts$get('youtube_output'): str(remedy::remedy_opts$get('youtube_output'))

remedy_opts$get('youtube_height'): str(remedy::remedy_opts$get('youtube_height'))

remedy_opts$get('youtube_width'): str(remedy::remedy_opts$get('youtube_width'))
}
\examples{
\dontrun{
remedy_example( 
    "https://www.youtube.com/watch?v=dQw4w9WgXcQ", 
    youtuber
    )
}
}
